Mudando para aceitar float

+++ apl1/src/Operando.java
@@ -1,23 +1,23 @@
 public class Operando extends Node {
-    private int operate;
+    private float operate;
 
     // Construtores
     public Operando() {
         super();
     }
 
-    public Operando(Node parent, int operate) {
+    public Operando(Node parent, float operate) {
         super(parent);
         this.operate = operate;
     }
 
     // Setter
-    public void setOperate(int operate) {
+    public void setOperate(float operate) {
         this.operate = operate;
     }
 
     // Método herdado da super classe que retorna o conteúdo do Node
-    @Override public Integer visit() {
+    @Override public Float visit() {
         return operate;
     }
 }
